requirements.txt
mysql-connector-python
tabulate
python-dotenv